requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Data handling
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter for analysis
jupyter>=1.0.0
notebook>=6.5.4

# Optional: OpenAI for advanced sentiment
openai>=1.0.0

# Data visualization
plotly>=5.14.0
